Move user authentication out of views, into controllers

+++ part-3/app/controllers/auctions_controller.rb
@@ -1,36 +1,56 @@
 get '/auctions/new' do
-  @auction = Auction.new
-  erb :'auctions/new'
+  if session[:user_id]
+    @auction = Auction.new
+    erb :'auctions/new'
+  else
+    redirect '/users/new'
+  end
 end
 
 post '/auctions' do
-  params[:auction][:user_id] = session[:user_id]
-  @auction = Auction.new(params[:auction])
-  if @auction.save
-    redirect "/users/#{session[:user_id]}"
+  if session[:user_id]
+    params[:auction][:user_id] = session[:user_id]
+    @auction = Auction.new(params[:auction])
+    if @auction.save
+      redirect "/users/#{session[:user_id]}"
+    else
+      erb :"auctions/new"
+    end
   else
-    erb :"auctions/new"
+    redirect '/users/new'
   end
 end
 
 get '/auctions/:id/edit' do
-  @auction = Auction.find(params[:id])
-  erb :'auctions/edit'
+  if session[:user_id]
+    @auction = Auction.find(params[:id])
+    erb :'auctions/edit'
+  else
+    redirect '/users/new'
+  end
 end
 
 put '/auctions/:id' do
-  @auction = Auction.find(params[:id])
-  @auction.update(params[:auction])
-  if @auction.save
-    redirect "/users/#{session[:user_id]}"
+  if session[:user_id]
+    @auction = Auction.find(params[:id])
+    @auction.update(params[:auction])
+    if @auction.save
+      redirect "/users/#{session[:user_id]}"
+    else
+      erb :"auctions/edit"
+    end
   else
-    erb :"auctions/edit"
+    redirect '/users/new'
   end
 end
 
 delete '/auctions/:id' do
-  auction = Auction.find(params[:id])
-  auction.destroy
+  if session[:user_id]
+    auction = Auction.find(params[:id])
+    auction.destroy
 
-  redirect "/users/#{session[:user_id]}"
+    redirect "/users/#{session[:user_id]}"
+  else
+    redirect '/users/new'
+  end
 end
